+++ client/src/utilities/format_date.js
@@ -1,0 +1,6 @@
+export const formatDateAsMonth = (date) => {
+    const year = date.getFullYear();
+    const monthZeroBased = date.getMonth();
+    const month = monthZeroBased < 9 ? '0' + (monthZeroBased + 1) : monthZeroBased + 1;
+    return `${year}-${month}`;
+};
